Can now be compiled using the Arduino IDE
An .ino file has been added, to allow compilation and flashing using the Arduino IDE. The main program has been extended to include a check if the EEPROM has been initialised. If not, the default CV values will be set in EEPROM.
Note that also "doProgramming()" has been uncommented, and the software version increased to 0x10. Although the safety decoder is working, additional tests and some cleaning up of the code may be necessary.